TST: Move helper modules out of ``_core/tests``
