+++ numpy/_core/_testutils_natype.py
