+++ numpy/_core/tests/test_multiarray.py
@@ -27,7 +27,7 @@ import numpy as np
 import numpy._core._multiarray_tests as _multiarray_tests
 from numpy._core._rational_tests import rational
 from numpy._core.multiarray import _get_ndarray_c_version, dot
-from numpy._core.tests._locales import CommaDecimalPointLocale
+from numpy._core._testutils_locales import CommaDecimalPointLocale
 from numpy.exceptions import AxisError, ComplexWarning
 from numpy.lib import stride_tricks
 from numpy.lib.recfunctions import repack_fields
